sample/requirements.txt
Django>=5.2,<6
djangorestframework>=3,<4
psycopg2-binary>=2,<3
requests>=2,<3
opentelemetry-api>=1.17,<2
opentelemetry-sdk>=1.17,<2
opentelemetry-instrumentation-django==0.59b0
opentelemetry-instrumentation-psycopg2==0.59b0
opentelemetry-instrumentation-logging==0.59b0
opentelemetry-distro==0.59b0
opentelemetry-distro[otlp]==0.59b0
opentelemetry-exporter-otlp>=1,<2
uwsgi>=2,<3